gateway/requirements.txt
fastapi
uvicorn[standard]
httpx
redis
python-dotenv

